Added SDK 10.7

+++ System/Library/Frameworks/WebKit.framework/Versions/A/Headers/DOMProgressEvent.h
@@ -32,10 +32,10 @@
 
 @interface DOMProgressEvent : DOMEvent
 @property(readonly) BOOL lengthComputable;
-@property(readonly) unsigned loaded;
-@property(readonly) unsigned total;
+@property(readonly) unsigned long long loaded;
+@property(readonly) unsigned long long total;
 
-- (void)initProgressEvent:(NSString *)typeArg canBubbleArg:(BOOL)canBubbleArg cancelableArg:(BOOL)cancelableArg lengthComputableArg:(BOOL)lengthComputableArg loadedArg:(unsigned)loadedArg totalArg:(unsigned)totalArg;
+- (void)initProgressEvent:(NSString *)typeArg canBubbleArg:(BOOL)canBubbleArg cancelableArg:(BOOL)cancelableArg lengthComputableArg:(BOOL)lengthComputableArg loadedArg:(unsigned long long)loadedArg totalArg:(unsigned long long)totalArg;
 @end
 
 #endif
